+++ src/main/java/io/github/pylonmc/pylon/base/PylonBase.java
@@ -1,8 +1,8 @@
 package io.github.pylonmc.pylon.base;
 
 import io.github.pylonmc.pylon.base.items.HealthTalisman;
+import io.github.pylonmc.pylon.base.items.tools.watering.Sprinkler;
 import io.github.pylonmc.pylon.base.misc.WaterCauldronRightClickRecipe;
-import io.github.pylonmc.pylon.base.items.watering.Sprinkler;
 import io.github.pylonmc.pylon.core.addon.PylonAddon;
 import lombok.Getter;
 import org.bukkit.Bukkit;
