+++ src/main/java/io/github/pylonmc/pylon/base/PylonBlocks.java
@@ -1,13 +1,8 @@
 package io.github.pylonmc.pylon.base;
 
-import io.github.pylonmc.pylon.base.items.EnrichedNetherrack;
-import io.github.pylonmc.pylon.base.items.Grindstone;
-import io.github.pylonmc.pylon.base.items.GrindstoneHandle;
-import io.github.pylonmc.pylon.base.items.MixingPot;
-import io.github.pylonmc.pylon.base.items.Pedestal;
-import io.github.pylonmc.pylon.base.items.MagicAltar;
+import io.github.pylonmc.pylon.base.items.multiblocks.*;
+import io.github.pylonmc.pylon.base.items.tools.watering.Sprinkler;
 import io.github.pylonmc.pylon.core.block.PylonBlockSchema;
-import io.github.pylonmc.pylon.base.items.watering.Sprinkler;
 import org.bukkit.Material;
 
 import static io.github.pylonmc.pylon.base.util.KeyUtils.pylonKey;
