+++ src/main/java/io/github/pylonmc/pylon/base/PylonEntities.java
@@ -1,7 +1,7 @@
 package io.github.pylonmc.pylon.base;
 
-import io.github.pylonmc.pylon.base.items.Grindstone;
-import io.github.pylonmc.pylon.base.items.Pedestal;
+import io.github.pylonmc.pylon.base.items.multiblocks.Grindstone;
+import io.github.pylonmc.pylon.base.items.multiblocks.Pedestal;
 import io.github.pylonmc.pylon.core.entity.PylonEntitySchema;
 import org.bukkit.entity.ItemDisplay;
