+++ src/main/java/io/github/pylonmc/pylon/base/items/weapons/RecoilArrow.java
@@ -1,4 +1,4 @@
-package io.github.pylonmc.pylon.base.items;
+package io.github.pylonmc.pylon.base.items.weapons;
 
 import com.destroystokyo.paper.event.player.PlayerReadyArrowEvent;
 import io.github.pylonmc.pylon.core.item.PylonItem;
